Don't rejoin messages after splitting
Joining records back together while batching sends
causes problems for multiline parsing.

For example, with the records

```
2020-10-05T10:00:00Z First line of an error message
  second line of an error message
2020-10-05T10:00:00Z A different error message
```

if multiline parsing is set correctly, we expect this
to be split into two records, the first
of two lines and the second of one line. However, if
records are joined at the point of sumo publishing, then
we may well get all three lines as a single record.

This is mostly a problem for multiline parsing as you can
treat each line as an individual message if multiline is
not used, but you cannot do that if needing to do the
multiline separation before sending to sumo.